transcription-service/requirements.txt
flask
azure-storage-blob
openai
openai-whisper
python-dotenv
